requirements.txt
mongoengine==0.17.0
Pillow==5.4.1